+++ govtool/frontend/src/components/organisms/ControlledField/Input.tsx
@@ -1,0 +1,32 @@
+import { useCallback } from "react";
+import { Controller, get } from "react-hook-form";
+
+import { Field } from "@molecules";
+
+import { ControlledInputProps, RenderInputProps } from "./types";
+
+export const Input = ({
+  control,
+  name,
+  errors,
+  rules,
+  ...props
+}: ControlledInputProps) => {
+  const errorMessage = get(errors, name)?.message as string;
+
+  const renderInput = useCallback(
+    ({ field }: RenderInputProps) => (
+      <Field.Input {...props} {...field} errorMessage={errorMessage} />
+    ),
+    [errorMessage, props]
+  );
+
+  return (
+    <Controller
+      name={name}
+      control={control}
+      rules={rules}
+      render={renderInput}
+    />
+  );
+};