+++ infra/terraform/modules/govtool-ec2/outputs.tf
@@ -1,0 +1,7 @@
+output "govtool_eip" {
+  value = aws_eip.govtool.public_ip
+}
+
+output "frontend_domain" {
+  value = var.app_env == "beta" ? aws_eip.govtool.public_ip : aws_route53_record.frontend[0].fqdn
+}